feat(minHeap): Implementing a min-heap data structure
A heap is a binary tree with both a shape property and a heap property. The shape property enforces
the tree to behave in a balanced way by defining each node to have two children unless the node is
in the very last level. The heap property ensures that each node is less than or equal to either of
its child nodes if it is a min-heap, and vice versa in case of a max-heap.